more mike changes

+++ util.go
@@ -1,13 +1,13 @@
 package fq
 
-func max(a, b uint64) uint64 {
+func max(a, b float64) float64 {
 	if a >= b {
 		return a
 	}
 	return b
 }
 
-func min(a, b uint64) uint64 {
+func min(a, b float64) float64 {
 	if a <= b {
 		return a
 	}
